[BUGFIX] Split RemoteObjectUpdateEvent to make it backwards compatible
