+++ Classes/Resource/Event/RemoteObjectUpdateEvent.php
@@ -17,22 +17,14 @@ namespace MaxServ\FalS3\Resource\Event;
  * The TYPO3 project - inspiring people to share!
  */
 
-use Aws\Result;
-use Aws\S3\S3Client;
-use MaxServ\FalS3\Driver\AmazonS3Driver;
-use MaxServ\FalS3\Utility\RemoteObjectUtility;
-use TYPO3\CMS\Core\Resource\AbstractFile;
+use MaxServ\FalS3\CacheControl\RemoteObjectUpdateCacheControl;
 use TYPO3\CMS\Core\Resource\Driver\DriverInterface;
 use TYPO3\CMS\Core\Resource\Event\AfterFileMetaDataCreatedEvent;
 use TYPO3\CMS\Core\Resource\Event\AfterFileMetaDataUpdatedEvent;
 use TYPO3\CMS\Core\Resource\Event\AfterFileProcessingEvent;
-use TYPO3\CMS\Core\Resource\File;
 use TYPO3\CMS\Core\Resource\FileInterface;
 use TYPO3\CMS\Core\Resource\ProcessedFile;
-use TYPO3\CMS\Core\Resource\ProcessedFileRepository;
-use TYPO3\CMS\Core\Resource\ResourceFactory;
 use TYPO3\CMS\Core\Resource\Service\FileProcessingService;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * Class RemoteObjectUpdateEvent
@@ -40,12 +32,22 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
 class RemoteObjectUpdateEvent
 {
     /**
+     * @var RemoteObjectUpdateCacheControl
+     */
+    protected $remoteObjectUpdateCacheControl;
+
+    public function __construct(RemoteObjectUpdateCacheControl $remoteObjectUpdateCacheControl)
+    {
+        $this->remoteObjectUpdateCacheControl = $remoteObjectUpdateCacheControl;
+    }
+
+    /**
      * PSR-14 Event listener
      * @param AfterFileMetaDataCreatedEvent $event
      */
     public function afterFileMetaDataCreated(AfterFileMetaDataCreatedEvent $event): void
     {
-        $this->updateRemoteCacheDirectiveForFile($event->getFileUid());
+        $this->remoteObjectUpdateCacheControl->updateRemoteCacheDirectiveForFile($event->getFileUid());
     }
 
     /**
@@ -54,7 +56,7 @@ class RemoteObjectUpdateEvent
      */
     public function afterFileMetaDataUpdated(AfterFileMetaDataUpdatedEvent $event): void
     {
-        $this->updateRemoteCacheDirectiveForFile($event->getFileUid());
+        $this->remoteObjectUpdateCacheControl->updateRemoteCacheDirectiveForFile($event->getFileUid());
     }
 
     /**
@@ -68,8 +70,8 @@ class RemoteObjectUpdateEvent
                 $event->getProcessedFile()->getIdentifier()
             );
 
-            if ($this->remoteObjectNeedsUpdate($fileInfo)) {
-                $this->updateCacheControlDirectivesForRemoteObject($event->getProcessedFile());
+            if ($this->remoteObjectUpdateCacheControl->remoteObjectNeedsUpdate($fileInfo)) {
+                $this->remoteObjectUpdateCacheControl->updateCacheControlDirectivesForRemoteObject($event->getProcessedFile());
             }
         }
     }
@@ -85,7 +87,7 @@ class RemoteObjectUpdateEvent
     {
         $fileUid = (int)($data['file'] ?? 0);
         if ($fileUid > 0) {
-            $this->updateRemoteCacheDirectiveForFile($fileUid);
+            $this->remoteObjectUpdateCacheControl->updateRemoteCacheDirectiveForFile($fileUid);
         }
 
         return [$data];
@@ -115,114 +117,8 @@ class RemoteObjectUpdateEvent
         array $configuration
     ): void {
         $fileInfo = $driver->getFileInfoByIdentifier($processedFile->getIdentifier());
-        if ($this->remoteObjectNeedsUpdate($fileInfo)) {
-            $this->updateCacheControlDirectivesForRemoteObject($processedFile);
-        }
-    }
-
-    /**
-     * @param array $fileInfo
-     * @return bool
-     */
-    protected function remoteObjectNeedsUpdate(array $fileInfo): bool
-    {
-        return array_key_exists('mtime', $fileInfo) && (int)$fileInfo['mtime'] > (time() - 30);
-    }
-
-    /**
-     * @param int $fileUid
-     */
-    protected function updateRemoteCacheDirectiveForFile(int $fileUid): void
-    {
-        try {
-            $file = GeneralUtility::makeInstance(ResourceFactory::class)->getFileObject($fileUid);
-        } catch (\Exception $e) {
-            return;
-        }
-
-        if (!$file instanceof File) {
-            return;
-        }
-
-        if ($file->getStorage()->getDriverType() !== AmazonS3Driver::DRIVER_KEY) {
-            return;
-        }
-
-        $this->updateCacheControlDirectivesForRemoteObject($file);
-        $processedFileRepository = GeneralUtility::makeInstance(ProcessedFileRepository::class);
-        if ($processedFileRepository instanceof ProcessedFileRepository) {
-            $processedFiles = $processedFileRepository->findAllByOriginalFile($file);
-            array_walk(
-                $processedFiles,
-                function (ProcessedFile $processedFile) {
-                    $this->updateCacheControlDirectivesForRemoteObject($processedFile);
-                }
-            );
-        }
-    }
-
-    /**
-     * @param AbstractFile $file
-     */
-    protected function updateCacheControlDirectivesForRemoteObject(AbstractFile $file): void
-    {
-        $currentResource = null;
-
-        $client = RemoteObjectUtility::resolveClientForStorage($file->getStorage());
-        $driverConfiguration = RemoteObjectUtility::resolveDriverConfigurationForStorage($file->getStorage());
-
-        $key = '';
-
-        if (array_key_exists('basePath', $driverConfiguration) && !empty($driverConfiguration['basePath'])) {
-            $key .= trim($driverConfiguration['basePath'], '/') . '/';
-        }
-
-        $key .= ltrim($file->getIdentifier(), '/');
-
-        if (
-            array_key_exists('bucket', $driverConfiguration)
-            && $client instanceof S3Client
-        ) {
-            try {
-                $currentResource = $client->headObject(
-                    [
-                        'Bucket' => $driverConfiguration['bucket'],
-                        'Key' => $key
-                    ]
-                );
-            } catch (\Exception $e) {
-                // fail silently if a file doesn't exist
-            }
-        }
-
-        if ($file instanceof ProcessedFile) {
-            $cacheControl = RemoteObjectUtility::resolveCacheControlDirectivesForFile(
-                $file->getOriginalFile(),
-                true
-            );
-        } else {
-            $cacheControl = RemoteObjectUtility::resolveCacheControlDirectivesForFile($file);
-        }
-
-        if (
-            !empty($cacheControl)
-            && $currentResource instanceof Result
-            && $currentResource->hasKey('Metadata')
-            && is_array($currentResource->get('Metadata'))
-            && $currentResource->hasKey('CacheControl')
-            && strcmp($currentResource->get('CacheControl'), $cacheControl) !== 0
-        ) {
-            $client->copyObject(
-                [
-                    'Bucket' => $driverConfiguration['bucket'],
-                    'CacheControl' => $cacheControl,
-                    'ContentType' => $currentResource->get('ContentType'),
-                    'CopySource' => $driverConfiguration['bucket'] . '/' . S3Client::encodeKey($key),
-                    'Key' => $key,
-                    'Metadata' => $currentResource->get('Metadata'),
-                    'MetadataDirective' => 'REPLACE'
-                ]
-            );
+        if ($this->remoteObjectUpdateCacheControl->remoteObjectNeedsUpdate($fileInfo)) {
+            $this->remoteObjectUpdateCacheControl->updateCacheControlDirectivesForRemoteObject($processedFile);
         }
     }
 }
